requirements.txt
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==1.7.1
idna==2.8
keyring==10.1
keyrings.alt==1.3
numpy==1.17.2
picamera==1.13
pyasn1==0.1.9
PyAudio==0.2.11
pycparser==2.19
pycrypto==2.6.1
pygobject==3.22.0
PySoundFile==0.9.0.post1
python-espeak==0.5
pyxdg==0.25
requests==2.22.0
RTIMULib==7.2.1
SecretStorage==2.3.1
six==1.12.0
urllib3==1.25.6
